YD-627 Processed review comment
The location for the assertion operations wasn't logical. Moved them now out of the lookup service to a dedicated assertion service.
